LMB-391: move logic to own component rangorde-input
