+++ app/components/verkiezingen/draft-mandataris-list.js
@@ -1,7 +1,7 @@
 import Component from '@glimmer/component';
+
 import { action } from '@ember/object';
 import { tracked } from '@glimmer/tracking';
-import { restartableTask, timeout } from 'ember-concurrency';
 
 export default class DraftMandatarisListComponent extends Component {
   @tracked isModalOpen = false;
@@ -27,19 +27,6 @@ export default class DraftMandatarisListComponent extends Component {
     this.editRangorde = mandataris.id;
     this.editRangordeValue = mandataris.rangorde;
   }
-
-  validateRangorde = restartableTask(async (event) => {
-    await timeout(200);
-
-    let possibleOrder = this.findOrderInString(event.target?.value);
-
-    if (!possibleOrder) {
-      console.warn(`Geen getal gevonden in input veld`);
-      return;
-    }
-
-    console.log({ possibleOrder });
-  });
 
   @action
   closeEditRangorde() {
@@ -79,46 +66,5 @@ export default class DraftMandatarisListComponent extends Component {
     this.mandataris.isBestuurlijkeAliasVan = person;
     await this.mandataris.save();
     this.closeModal();
-  }
-
-  findOrderInString(possibleString) {
-    if (!possibleString) {
-      return null;
-    }
-
-    const firstWord = this.findFirstWordOfString(possibleString);
-    if (!firstWord) {
-      return null;
-    }
-
-    const lowercaseWord = firstWord.input?.toLowerCase();
-    if (Object.keys(this.rangordeAsStringMapping).includes(lowercaseWord)) {
-      return this.rangordeAsStringMapping[lowercaseWord];
-    } else {
-      return parseInt(lowercaseWord);
-    }
-  }
-
-  findFirstWordOfString(string) {
-    // eslint-disable-next-line no-useless-escape
-    const regex = new RegExp(/^([\w\-]+)/);
-    if (regex.test(string)) {
-      return string.match(regex);
-    }
-    return null;
-  }
-
-  get rangordeAsStringMapping() {
-    return {
-      eerste: 1,
-      tweede: 2,
-      derde: 3,
-      vierde: 4,
-      vijfde: 5,
-      zesde: 6,
-      zevende: 7,
-      achtste: 8,
-      negende: 9,
-    };
   }
 }
